+++ csrc/single_prefill.cu
@@ -35,8 +35,8 @@ using namespace flashinfer;
 
 void single_prefill_with_kv_cache(at::Tensor q, at::Tensor k, at::Tensor v, at::Tensor tmp,
                                   at::Tensor o, std::optional<at::Tensor> maybe_lse,
-                                  unsigned int mask_mode_code, unsigned int layout,
-                                  int32_t window_left ADDITIONAL_FUNC_PARAMS, int64_t cuda_stream) {
+                                  int64_t mask_mode_code, int64_t layout,
+                                  int64_t window_left ADDITIONAL_FUNC_PARAMS, int64_t cuda_stream) {
   auto device = q.device();
   unsigned int head_dim_qk = q.size(2);
   unsigned int kv_len, qo_len, num_kv_heads, num_qo_heads;
